refactor(web): extract session services into pi-backend
